+++ pkg/omreport/omreport.go
@@ -344,7 +344,7 @@ func (or *OMReport) StorageVdisk() ([]Value, error) {
 			},
 		})
 
-		if len(fields) > 5 {
+		if len(fields) > 17 {
 			values = append(values, Value{
 				Name:  "storage_vdisk_raidlevel",
 				Value: getNumberFromString(fields[6]),
@@ -378,6 +378,46 @@ func (or *OMReport) StorageVdisk() ([]Value, error) {
 			values = append(values, Value{
 				Name:  "storage_vdisk_cache_policy",
 				Value: vdiskCachePolicy(fields[15]),
+				Labels: map[string]string{
+					"vdisk":             id,
+					"vdisk_name":        fields[2],
+					controllerNameLabel: controllerName,
+				},
+			})
+		} else {
+			values = append(values, Value{
+				Name:  "storage_vdisk_raidlevel",
+				Value: getNumberFromString(fields[5]),
+				Labels: map[string]string{
+					"vdisk":             id,
+					"vdisk_name":        fields[2],
+					controllerNameLabel: controllerName,
+				},
+			})
+
+			values = append(values, Value{
+				Name:  "storage_vdisk_read_policy",
+				Value: vdiskReadPolicy(fields[12]),
+				Labels: map[string]string{
+					"vdisk":             id,
+					"vdisk_name":        fields[2],
+					controllerNameLabel: controllerName,
+				},
+			})
+
+			values = append(values, Value{
+				Name:  "storage_vdisk_write_policy",
+				Value: vdiskWritePolicy(fields[13]),
+				Labels: map[string]string{
+					"vdisk":             id,
+					"vdisk_name":        fields[2],
+					controllerNameLabel: controllerName,
+				},
+			})
+
+			values = append(values, Value{
+				Name:  "storage_vdisk_cache_policy",
+				Value: vdiskCachePolicy(fields[14]),
 				Labels: map[string]string{
 					"vdisk":             id,
 					"vdisk_name":        fields[2],
@@ -451,7 +491,7 @@ func (or *OMReport) Nics(nicList ...string) ([]Value, error) {
 		id := strings.Replace(fields[0], ":", "_", -1)
 		ts := map[string]string{"id": id, "device": fields[1]}
 		var ret string
-		if fields[4] == "Connected" || fields[4] == "Full" {
+		if fields[4] == "Connected" || fields[4] == "Full" || fields[4] == "Not Applicable" {
 			ret = "0"
 		} else {
 			ret = "1"